requirements.txt
PyYAML~=5.3.1
simpleobsws~=0.0.7
octorest~=0.4
setuptools~=50.3.2